feat: Add MapScreen.hava and simple map implementation
Add a map test menu to the title screen.
You can move a ship within a map, and the color distinction of what you clear on the map, what you can go, what you can't go.
The grid size of the map can be modified in MapScreen.java, and the cell size and margin can be modified in DrawManager.java.